+++ cognite/powerops/__init__.py
@@ -1,5 +1,5 @@
-from cognite.powerops.case import Case
 from cognite.powerops.client import PowerOpsClient
+from cognite.powerops.client.data_classes.case import Case
 from cognite.powerops.version import __version__
 
 __all__ = ["__version__", "Case", "PowerOpsClient"]