+++ CHANGELOG.md
@@ -13,11 +13,16 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
-## [0.12.0] - 19-06-23
+## [0.13.0] - 19-06-23
 ### Changed
 
 * Removed unnecessary dependencies
 
+## [0.12.0] - 08-06-23
+### Added
+
+* Ability to trigger SHOP runs via CogShop.
+* Retrieval of ShopRun results.
 
 ## [0.11.0] - 08-06-23
 ### Feature